Changed db engine to MySql2

+++ Blog/db/schema.rb
@@ -23,7 +23,7 @@ ActiveRecord::Schema.define(version: 2020_06_11_123923) do
   create_table "comments", force: :cascade do |t|
     t.string "commenter"
     t.text "body"
-    t.integer "article_id", null: false
+    t.bigint "article_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["article_id"], name: "index_comments_on_article_id"
